Make backend/frontend builds explicit; isolate e2e tmux; split integration tests
Three related testing-policy cleanups.

1. Decouple the frontend build from the Rust compile (drop WEAVER_SKIP_FRONTEND)
   The SPA is served from static/dist at runtime (ServeDir), not embedded in the
   binary, so building it never needs to be part of `cargo build`. The old
   policy built it on *every* build and told everyone to set WEAVER_SKIP_FRONTEND
   to opt out — a backwards default. Invert it: build.rs builds the SPA only
   under the new `embed-frontend` cargo feature, writing a placeholder otherwise.
   - Backend: `cargo build` / `cargo test` — fast, no Node, no env var.
   - Frontend bundle: `cargo build --features embed-frontend` (or
     `npm --prefix crates/loom/frontend run build`).
   - Frontend tests: the Playwright `e2e/` suite, which builds the SPA itself.
   CI `build` job uses the feature (still catches frontend dep breakage);
   `lint`/`test` drop the env var. README + AGENTS.md document the split.

2. Isolate the e2e tmux socket (#22)
   The Playwright fixture drove a real `loom serve` on the machine-global default
   tmux server — the same one the user's live weaver-<id> agents run on, so its
   teardown could disturb them. It now pins a private `WEAVER_TMUX_SOCKET`
   (`tmux -L weaver-e2e-<pid>-<rand>`) per run, pre-cleaned at startup and reaped
   on teardown, exactly like the Rust harness. Every session it creates/kills is
   scoped to that socket; the default socket is untouched.

3. Split the 858-line tests/integration.rs into tests/integration/
   A shared `fixtures::TestServer` harness (isolated WEAVER_HOME/db, private tmux
   socket, throwaway git repo, terminal-WS helpers) plus focused, `#[serial]`
   cases: sessions, terminal, scratch, files, branches, archive. Only the PTY
   `terminal` test keeps the CI skip, so the rest of the coverage now runs in CI
   (the monolith skipped everything under CI). The hook→status path stays in
   tests/hook_monitor.rs. Adds the `serial_test` dev-dep.
