+++ crates/loom/build.rs
@@ -10,26 +10,31 @@ fn mtime(path: &Path) -> SystemTime {
         .unwrap_or(SystemTime::UNIX_EPOCH)
 }
 
-// Builds the Vue frontend into `static/dist`. Skipped (with a placeholder page
-// written instead) when `WEAVER_SKIP_FRONTEND` is set, npm is unavailable, or
-// the frontend sources do not exist yet — so the backend can be built/tested
-// without a Node toolchain.
+// Builds the Vue frontend into `static/dist`, but only under the `embed-frontend`
+// cargo feature (and when npm + the frontend sources are present). The SPA is
+// served from `static/dist` at runtime (see `web::static_dir`), not embedded in
+// the binary, so the bundle is independent of the Rust compile: a plain
+// `cargo build` / `cargo test` skips it — the fast, Node-free path for backend
+// work — and writes a placeholder page instead. Build a UI-bearing tree with
+// `cargo build --features embed-frontend`, or run the SPA build directly with
+// `npm --prefix crates/loom/frontend run build`.
 fn main() {
     // Every file that feeds the frontend build: changing any of them reruns
     // this script (and therefore rspack). `frontend/src` covers the Vue/TS
-    // sources and the HTML template; the rest are build-config inputs.
+    // sources and the HTML template; the rest are build-config inputs. (Cargo
+    // already reruns build scripts when the active feature set changes, so the
+    // `embed-frontend` toggle needs no explicit rerun directive.)
     println!("cargo:rerun-if-changed=frontend/src");
     println!("cargo:rerun-if-changed=frontend/package.json");
     println!("cargo:rerun-if-changed=frontend/package-lock.json");
     println!("cargo:rerun-if-changed=frontend/rspack.config.js");
     println!("cargo:rerun-if-changed=frontend/postcss.config.mjs");
     println!("cargo:rerun-if-changed=frontend/tsconfig.json");
-    println!("cargo:rerun-if-env-changed=WEAVER_SKIP_FRONTEND");
 
     let dist = Path::new("static/dist");
     let frontend = Path::new("frontend");
 
-    let skip = std::env::var("WEAVER_SKIP_FRONTEND").is_ok();
+    let embed = std::env::var_os("CARGO_FEATURE_EMBED_FRONTEND").is_some();
     let have_npm = Command::new("npm")
         .arg("--version")
         .stdout(std::process::Stdio::null())
@@ -39,7 +44,7 @@ fn main() {
         .unwrap_or(false);
     let have_sources = frontend.join("src/main.ts").exists();
 
-    if skip || !have_npm || !have_sources {
+    if !embed || !have_npm || !have_sources {
         std::fs::create_dir_all(dist).ok();
         let index = dist.join("index.html");
         if !index.exists() {
@@ -47,8 +52,10 @@ fn main() {
                 &index,
                 "<!doctype html><meta charset=utf-8><title>weaver</title>\
                  <body style=\"font-family:sans-serif;padding:2rem\">\
-                 <h1>weaver</h1><p>Frontend not built. \
-                 Rebuild with npm available and <code>WEAVER_SKIP_FRONTEND</code> unset.</p>",
+                 <h1>weaver</h1><p>Frontend not built. Build it with \
+                 <code>cargo build --features embed-frontend</code> \
+                 (needs Node + npm), or \
+                 <code>npm --prefix crates/loom/frontend run build</code>.</p>",
             )
             .ok();
         }
