+++ crates/loom/tests/integration/fixtures.rs
@@ -1,0 +1,197 @@
+//! Shared harness for the loom integration suites: a real server bound to a
+//! random port, backed by an isolated weaver home, a private tmux server, and a
+//! throwaway git repo. See the sibling modules (`sessions`, `terminal`,
+//! `scratch`, `files`, `branches`, `archive`) for the focused cases.
+//!
+//! `TestServer::start` mutates process-global env (`WEAVER_HOME` /
+//! `WEAVER_API` / `WEAVER_TMUX_SOCKET`), so every test that uses it is marked
+//! `#[serial]` — they share one binary and would otherwise race on that env.
+
+use std::net::SocketAddr;
+use std::path::Path;
+use std::process::Command;
+use std::time::Duration;
+
+use futures_util::SinkExt;
+use loom::client::Client;
+use loom::events::EventBus;
+use loom::web::AppState;
+use loom::{db, server};
+use tempfile::TempDir;
+use tokio::net::TcpListener;
+use tokio_tungstenite::tungstenite::Message;
+
+/// Run `program args` in `dir`, asserting it succeeds.
+pub fn sh(dir: &Path, program: &str, args: &[&str]) {
+    let status = Command::new(program)
+        .args(args)
+        .current_dir(dir)
+        .status()
+        .unwrap_or_else(|e| panic!("failed to run {program}: {e}"));
+    assert!(status.success(), "{program} {args:?} failed");
+}
+
+/// Pins tmux to a throwaway server (`tmux -L <name>`) for the test and kills it
+/// on drop — so the suite can never see or disturb the user's real sessions,
+/// even if the test panics. See `loom::tmux::socket_args`.
+struct TmuxSocket(String);
+
+impl TmuxSocket {
+    fn install() -> Self {
+        let name = format!("weaver-test-{}", std::process::id());
+        std::env::set_var("WEAVER_TMUX_SOCKET", &name);
+        // Clear any stale server left by a crashed prior run with this pid.
+        let _ = Command::new("tmux")
+            .args(["-L", &name, "kill-server"])
+            .status();
+        TmuxSocket(name)
+    }
+}
+
+impl Drop for TmuxSocket {
+    fn drop(&mut self) {
+        let _ = Command::new("tmux")
+            .args(["-L", &self.0, "kill-server"])
+            .status();
+    }
+}
+
+/// Build a throwaway git repo with a single commit on `main`.
+fn init_repo(dir: &Path) {
+    sh(dir, "git", &["init", "-b", "main"]);
+    sh(dir, "git", &["config", "user.email", "t@t.test"]);
+    sh(dir, "git", &["config", "user.name", "Test"]);
+    std::fs::write(dir.join("README.md"), "hello\n").unwrap();
+    sh(dir, "git", &["add", "."]);
+    sh(dir, "git", &["commit", "-m", "init"]);
+}
+
+/// A running loom server with fully isolated state: its own temp `WEAVER_HOME`
+/// (and therefore its own sqlite db), a private tmux server, and a throwaway git
+/// repo. Drop tears tmux down; the temp dirs clean themselves up.
+pub struct TestServer {
+    pub client: Client,
+    pub addr: SocketAddr,
+    repo: TempDir,
+    _home: TempDir,
+    _tmux: TmuxSocket,
+}
+
+impl TestServer {
+    /// Boot a server on a random port with isolated state and wait for it to
+    /// answer `/api/health`. The caller must be `#[serial]`: setup writes
+    /// process-global env.
+    pub async fn start() -> Self {
+        // Isolate weaver state in a temp home (its own db) and tmux on a private
+        // server for the lifetime of the test.
+        let home = tempfile::tempdir().unwrap();
+        std::env::set_var("WEAVER_HOME", home.path());
+        let tmux = TmuxSocket::install();
+
+        let repo = tempfile::tempdir().unwrap();
+        init_repo(repo.path());
+
+        let listener = TcpListener::bind("127.0.0.1:0").await.unwrap();
+        let addr = listener.local_addr().unwrap();
+        let pool = db::connect(&db::default_db_path()).await.unwrap();
+        let state = AppState {
+            db: pool,
+            bus: EventBus::new(),
+            addr: addr.to_string(),
+        };
+        tokio::spawn(server::serve(state, listener));
+
+        std::env::set_var("WEAVER_API", format!("http://{addr}"));
+        let client = Client::new();
+        for _ in 0..60 {
+            if client.get("/api/health").await.is_ok() {
+                break;
+            }
+            tokio::time::sleep(Duration::from_millis(50)).await;
+        }
+
+        Self {
+            client,
+            addr,
+            repo,
+            _home: home,
+            _tmux: tmux,
+        }
+    }
+
+    /// Path to the throwaway repo used as `cwd` when creating sessions.
+    pub fn repo_path(&self) -> &Path {
+        self.repo.path()
+    }
+
+    /// `repo_path()` as the lossy string the API expects in `cwd`.
+    pub fn cwd(&self) -> String {
+        self.repo.path().to_string_lossy().into_owned()
+    }
+}
+
+// ---------------------------------------------------------------------------
+// Terminal WebSocket helpers (used by `terminal.rs`)
+// ---------------------------------------------------------------------------
+
+pub type TermWs =
+    tokio_tungstenite::WebSocketStream<tokio_tungstenite::MaybeTlsStream<tokio::net::TcpStream>>;
+
+/// Connect a terminal WebSocket to a session. No `Origin` header is sent, so the
+/// server's same-origin check takes the missing-Origin (non-browser) path.
+pub async fn connect_terminal(addr: &SocketAddr, id: &str) -> TermWs {
+    let url = format!("ws://{addr}/api/sessions/{id}/terminal");
+    let (ws, _resp) = tokio_tungstenite::connect_async(url)
+        .await
+        .expect("terminal websocket should connect");
+    ws
+}
+
+/// A `0x00`-prefixed keystroke frame.
+pub fn input_frame(s: &str) -> Vec<u8> {
+    let mut v = vec![0x00u8];
+    v.extend_from_slice(s.as_bytes());
+    v
+}
+
+/// A `0x01 <cols_be> <rows_be>` resize frame.
+pub fn resize_frame(cols: u16, rows: u16) -> Vec<u8> {
+    let mut v = vec![0x01u8];
+    v.extend_from_slice(&cols.to_be_bytes());
+    v.extend_from_slice(&rows.to_be_bytes());
+    v
+}
+
+/// Send a keystroke frame down a terminal WebSocket.
+pub async fn send_input(ws: &mut TermWs, s: &str) {
+    ws.send(Message::Binary(input_frame(s).into()))
+        .await
+        .unwrap();
+}
+
+/// Accumulate ALL binary output frames into one buffer (the marker may span
+/// frames and is interleaved with ANSI escapes) until `marker` appears or the
+/// timeout elapses. Returns the decoded buffer either way.
+pub async fn drain_until(ws: &mut TermWs, marker: &str, timeout: Duration) -> String {
+    use futures_util::StreamExt;
+    let mut buf: Vec<u8> = Vec::new();
+    let deadline = tokio::time::Instant::now() + timeout;
+    loop {
+        let remaining = deadline.saturating_duration_since(tokio::time::Instant::now());
+        if remaining.is_zero() {
+            break;
+        }
+        match tokio::time::timeout(remaining, ws.next()).await {
+            Ok(Some(Ok(Message::Binary(b)))) => {
+                buf.extend_from_slice(&b);
+                if String::from_utf8_lossy(&buf).contains(marker) {
+                    break;
+                }
+            }
+            Ok(Some(Ok(_))) => {}                 // text/ping/pong/close
+            Ok(Some(Err(_))) | Ok(None) => break, // stream error / end
+            Err(_) => break,                      // timeout
+        }
+    }
+    String::from_utf8_lossy(&buf).to_string()
+}
